# Speed up Time.zone.now